fix(coding-agent): retry provider stream stalls on the shared budget
Provider stream watchdog failures ("Provider stream start timed out after
<n>ms" and "Idle timeout waiting for provider stream after <n>ms") were
special-cased: the second consecutive stall escalated straight to the
fallback chain, bypassing settings.retry.maxRetries. Without a configured
chain the turn surrendered after a single same-model probe and surfaced
"Retry failed after 1 attempts".

Stalls are ordinary transient failures (isRetryableErrorMessage already
classifies them so). They now consume the same bounded same-model budget as
every other retryable class and escalate to the chain only on exhaustion.
